lib/classifier/pyproject.toml
[project]
name = "classifier"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic>=2.12.5",
    "tqdm>=4.67.3",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "ruamel-yaml>=0.19.1",
]


